Do 2019 Day 9 Part 1

+++ 2019/day/2/IntcodeAddition.cs
@@ -6,9 +6,9 @@ namespace AdventOfCode {
     class IntcodeAddition : IntcodeInstruction {
         public IntcodeAddition() : base(1, 3) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
-            int x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
-            int y = IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
+            long x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
+            long y = IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
             IntcodeInstruction.SetValue(2, computer, parameters, parameterModes, x + y);
             return true;
         }
