+++ 2019/day/2/IntcodeHalt.cs
@@ -6,7 +6,7 @@ namespace AdventOfCode {
     class IntcodeHalt : IntcodeInstruction {
         public IntcodeHalt() : base(99, 0) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
             computer.Halt();
             return false;
         }
