+++ 2019/day/5/IntcodeInput.cs
@@ -6,8 +6,8 @@ namespace AdventOfCode {
     class IntcodeInput : IntcodeInstruction {
         public IntcodeInput() : base(3, 1) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
-            int input = computer.GetInput();
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
+            long input = computer.GetInput();
             if (!computer.isRunning()) return false;
             
             IntcodeInstruction.SetValue(0, computer, parameters, parameterModes, input);
