+++ 2019/day/5/IntcodeJumpIfTrue.cs
@@ -6,11 +6,11 @@ namespace AdventOfCode {
     class IntcodeJumpIfTrue : IntcodeInstruction {
         public IntcodeJumpIfTrue() : base(5, 2) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
-            int x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
+            long x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
 
             if (x != 0) {
-                int instructionPointer = IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
+                int instructionPointer = (int) IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
                 computer.SetInstructionPointer(instructionPointer);
                 return false;
             }
