+++ 2019/day/5/IntcodeLessThan.cs
@@ -6,10 +6,10 @@ namespace AdventOfCode {
     class IntcodeLessThan : IntcodeInstruction {
         public IntcodeLessThan() : base(7, 3) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
-            int x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
-            int y = IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
-            IntcodeInstruction.SetValue(2, computer, parameters, parameterModes, Convert.ToInt32(x < y));
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
+            long x = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
+            long y = IntcodeInstruction.GetValue(1, computer, parameters, parameterModes);
+            IntcodeInstruction.SetValue(2, computer, parameters, parameterModes, Convert.ToInt64(x < y));
             return true;
         }
     }
