+++ 2019/day/5/IntcodePrintOutput.cs
@@ -6,8 +6,8 @@ namespace AdventOfCode {
     class IntcodePrintOutput : IntcodeInstruction {
         public IntcodePrintOutput() : base(4, 1) {}
 
-        override public bool Execute(IntcodeComputer computer, int[] parameters, int[] parameterModes) {
-            int output = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
+        override public bool Execute(IntcodeComputer computer, long[] parameters, int[] parameterModes) {
+            long output = IntcodeInstruction.GetValue(0, computer, parameters, parameterModes);
             computer.AddOutput(output);
             Console.WriteLine("Diagnostic code: " + output);
             return true;
